-update db, update api(updateBy, InsertBy)

+++ PetRescue/PetRescue.Data/Models/AdoptionRegisterForm.cs
@@ -12,14 +12,14 @@ namespace PetRescue.Data.Models
         public string Email { get; set; }
         public string Job { get; set; }
         public string Address { get; set; }
-        public int HouseType { get; set; }
-        public int FrequencyAtHome { get; set; }
+        public byte HouseType { get; set; }
+        public byte FrequencyAtHome { get; set; }
         public bool HaveChildren { get; set; }
-        public int? ChildAge { get; set; }
+        public byte? ChildAge { get; set; }
         public bool BeViolentTendencies { get; set; }
         public bool HaveAgreement { get; set; }
-        public int HavePet { get; set; }
-        public int AdoptionRegisterStatus { get; set; }
+        public byte HavePet { get; set; }
+        public byte AdoptionRegisterStatus { get; set; }
         public Guid InsertedBy { get; set; }
         public DateTime InsertedAt { get; set; }
         public Guid? UpdatedBy { get; set; }
