+++ PetRescue/PetRescue.Data/Models/RescueReport.cs
@@ -6,8 +6,8 @@ namespace PetRescue.Data.Models
     public partial class RescueReport
     {
         public Guid RescueReportId { get; set; }
-        public int PetAttribute { get; set; }
-        public int ReportStatus { get; set; }
+        public byte PetAttribute { get; set; }
+        public byte ReportStatus { get; set; }
         public Guid InsertedBy { get; set; }
         public DateTime InsertedAt { get; set; }
         public Guid? UpdatedBy { get; set; }
